+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * 【core】         FileUtil.mainName修正处理逻辑
 * 【core】         CharUtil增加isFileSeparator方法
 * 【core】         增加UUID类，提升Simple模式下性能
+* 【poi】           ExcelUtil增加setStyleSet方法，修改write逻辑，对于单列数据输出，而非忽略（感谢@【宁波】mojie126）
 
 ### Bug修复